Rules
You will be given a random word, and the game tells you the length of the word, as well as the number of lives that you have. Your end goal is to guess the word correctly before your lives run out.

When you guess the correct letter in the correct position of the word, then it will count the number of BULLS you have. When you guess the correct letter but it is in the wrong position then the game will tell you the number of COWS you have. For a win, you have to have the number of bulls equivalent to the length of the word.